% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{barcode}
\alias{barcode}
\title{Generate JsBarcode}
\usage{
barcode(id, code)
}
\arguments{
\item{id}{ID for SVG html element}

\item{code}{Value to convert to barcode}
}
\description{
Generate JsBarcode
}
\note{
This is used mainly in \code{render_receipt}
}
