% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include.R
\name{includeJsBarcode}
\alias{includeJsBarcode}
\title{Include JsBarcode JS Script}
\usage{
includeJsBarcode(cdn = FALSE)
}
\arguments{
\item{cdn}{Indicate whether to get JsBarcode from CDN or server}
}
\description{
Include JsBarcode JS Script
}
