% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_receipt}
\alias{render_receipt}
\title{Render Receipt as HTML}
\usage{
render_receipt(logo, transactionId, cart, name, template = system.file(package
  = "CannaBarcode", "templates", "template.rmd"))
}
\arguments{
\item{logo}{Client logo}

\item{transactionId}{ID of transaction}

\item{cart}{table of items to print}

\item{name}{Name of user}

\item{template}{rmd template}
}
\description{
Render Receipt as HTML
}
\note{
For use in \href{https://www.cannadata.com}{CannaData}
}
