% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{renderBarcode}
\alias{renderBarcode}
\alias{barcodeOutput}
\title{Render barcodes in Shiny applications}
\usage{
renderBarcode(expr, env = parent.frame(), quoted = FALSE, ...)

barcodeOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{expr}{An expression that returns a list.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}

\item{...}{additional options passed to JsBarcode}

\item{outputId}{output variable to read the plot/image from.}

\item{width}{Image width/height. Must be a valid CSS unit (like
\code{"100\%"}, \code{"400px"}, \code{"auto"}) or a number, which will be
coerced to a string and have \code{"px"} appended. These two arguments are
ignored when \code{inline = TRUE}, in which case the width/height of a plot
must be specified in \code{renderPlot()}. Note that, for height, using
\code{"auto"} or \code{"100\%"} generally will not work as expected,
because of how height is computed with HTML/CSS.}

\item{height}{Image width/height. Must be a valid CSS unit (like
\code{"100\%"}, \code{"400px"}, \code{"auto"}) or a number, which will be
coerced to a string and have \code{"px"} appended. These two arguments are
ignored when \code{inline = TRUE}, in which case the width/height of a plot
must be specified in \code{renderPlot()}. Note that, for height, using
\code{"auto"} or \code{"100\%"} generally will not work as expected,
because of how height is computed with HTML/CSS.}
}
\description{
Render barcodes in Shiny applications
}
